smb: Update the configuration logic for smbd ports
This change lets us switch the default listening ports (445 and 139)
for smbd to a custom configured port or to 445 when host networking is
disabled. Previously, with host networking off, the legacy port 139 was
never published even though the daemon was listening on it inside the
container. For consistency, regardless of whether a custom SMB port is
set, the `smb ports` parameter in smb.conf is used. Since the SMB daemon
publishes whatever port is defined in the configuration, it makes sense
to remove the checks for clustered and custom ports while adding the
`smb ports` parameter.

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>
Resolves: IBMCEPH-13332 (prereq)
(cherry picked from commit 82b299cf05f883b43b977a2c7fe9161f756f2ea4)